Criação pasta components (contém produtos e modal)
Objetivo de reduzir o código da lista
