+++ flor/src/pages/components/Produto/Produto.css
@@ -34,7 +34,6 @@
 }
 
 
-
 .elementosProd{
     display: flex;
     flex-direction: row;
