+++ flor/src/pages/components/ProdutoModal/ProdutoModal.css
@@ -72,12 +72,6 @@
   position:relative;
   bottom:0;
 }
-.accordionModal{
-  flex-direction: column;
-  flex-grow:1;
-  align-items: center;
-  justify-content: center;
-}
 
 .comentarioModal{
     font-weight: 300;
